+++ commons-civil/src/main/java/uk/gov/hmcts/reform/civil/launchdarkly/FeatureToggle.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.aspect;
+package uk.gov.hmcts.reform.civil.launchdarkly;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
