+++ commons-civil/src/main/java/uk/gov/hmcts/reform/civil/launchdarkly/FeatureToggleAspect.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.aspect;
+package uk.gov.hmcts.reform.civil.launchdarkly;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -6,7 +6,6 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.civil.launchdarkly.FeatureToggleService;
 
 @Slf4j
 @Aspect
