+++ commons-civil/src/test/java/uk/gov/hmcts/reform/civil/launchdarkly/FeatureToggleAspectTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.aspect;
+package uk.gov.hmcts.reform.civil.launchdarkly;
 
 import com.launchdarkly.sdk.LDUser;
 import com.launchdarkly.sdk.server.LDClient;
@@ -13,6 +13,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import uk.gov.hmcts.reform.civil.launchdarkly.FeatureToggle;
+import uk.gov.hmcts.reform.civil.launchdarkly.FeatureToggleAspect;
 import uk.gov.hmcts.reform.civil.launchdarkly.FeatureToggleService;
 
 import static org.mockito.ArgumentMatchers.any;
