+++ commons-civil/src/test/java/uk/gov/hmcts/reform/civil/sendgrid/config/SendGridAutoConfigurationTests.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sendgrid.config;
+package uk.gov.hmcts.reform.civil.sendgrid.config;
 
 import com.sendgrid.SendGrid;
 import org.junit.jupiter.api.AfterEach;
